+++ DeviceAIRSPY.cpp
@@ -161,7 +161,7 @@ namespace Device {
 
 	void AIRSPY::callback(CFLOAT32* data, int len)
 	{
-		RAW r = { Format::CF32 , data, len * sizeof(CFLOAT32) };
+		RAW r = { Format::CF32 , data, (int)(len * sizeof(CFLOAT32)) };
 		Send(&r, 1 );
 	}
