+++ DeviceFileRAW.cpp
@@ -104,7 +104,7 @@ namespace Device {
 
 		if (layout != FileLayout::Stereo) throw "FILE RAW: layout not implemented.";
 
-		RAW r = { format, buffer.data(), buffer.size() };
+		RAW r = { format, buffer.data(), (int)buffer.size() };
 		Send(&r, 1);
 
 		if (file->eof()) Stop();
